Fixed warnings.

+++ Source/iOS/UIImage+ECCore.m
@@ -134,8 +134,8 @@
 	
     // Build a context that's the same dimensions as the new size
     CGContextRef bitmap = CGBitmapContextCreate(NULL,
-                                                newRect.size.width,
-                                                newRect.size.height,
+                                                (size_t) newRect.size.width,
+                                                (size_t) newRect.size.height,
                                                 bitsPerComponent,
                                                 0,
                                                 colorSpace,
@@ -210,14 +210,14 @@
 
 - (UIImage*)imageAsGreyscale:(CGFloat)scale
 {
-	int width = self.size.width;
-	int height = self.size.height;
+	CGFloat width = self.size.width;
+	CGFloat height = self.size.height;
 	
 	CGColorSpaceRef colorSpace = CGColorSpaceCreateDeviceGray();
 	
 	CGContextRef context = CGBitmapContextCreate (nil,
-												  width,
-												  height,
+												  (size_t) width,
+												  (size_t) height,
 												  8,      // bits per component
 												  0,
 												  colorSpace,
